upload to github

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: hayashikdi <hayashikdi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/18 14:58:30 by hayashikdi        #+#    #+#             */
-/*   Updated: 2023/02/10 14:42:17 by hayashikdi       ###   ########.fr       */
+/*   Updated: 2023/03/04 20:21:03 by hayashikdi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	if (!s || !(*f))
 		return ;
